Sort URI Handlers by length when adding, so that URI matching always looks for the closest match
By sorting the URI handlers based on the length of the URI, longest first, we ensure that requests that do not match exactly
match the nearest uri instead of the first, regardless of the order in which they where added.

E.g.

If we have this setup:

/A - AHandler
/A/B - ABHandler

A request to /A/B/C will be handled by the ABHandler, regardless of the order in which they where added.